# IHM_VRAI_CUSTOM_NET/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebGuitare",
    "version": "1.0"
  },
  "paths": {
    "/api/Authentification/GetAuthentificationByEmail": {
      "get": {
        "tags": [
          "Authentification"
        ],
        "operationId": "GetAuthentificationByEmail",
        "parameters": [
          {
            "name": "P_Email",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_AUTHENTIFICATION"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_AUTHENTIFICATION"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_AUTHENTIFICATION"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Authentification/GetAuthentificationById": {
      "get": {
        "tags": [
          "Authentification"
        ],
        "operationId": "GetAuthentificationById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_AUTHENTIFICATION"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_AUTHENTIFICATION"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_AUTHENTIFICATION"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authentification/AddAuthentification": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "operationId": "AddAuthentification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authentification/VerifAuthentification": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "operationId": "VerifAuthentification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authentification/HashPassword": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "operationId": "HashPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_HASHMDP"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_HASHMDP"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_HASHMDP"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authentification/UpdateAuthentification": {
      "put": {
        "tags": [
          "Authentification"
        ],
        "operationId": "UpdateAuthentification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_AUTHENTIFICATION"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authentification/DeleteAuthentification": {
      "delete": {
        "tags": [
          "Authentification"
        ],
        "operationId": "DeleteAuthentification",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/BoisType/GetBoisTypeById": {
      "get": {
        "tags": [
          "BoisType"
        ],
        "operationId": "GetBoisTypeById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_BOISTYPES"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_BOISTYPES"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_BOISTYPES"
                }
              }
            }
          }
        }
      }
    },
    "/api/BoisType/GetBoisTypeByNom": {
      "get": {
        "tags": [
          "BoisType"
        ],
        "operationId": "GetBoisTypeByNom",
        "parameters": [
          {
            "name": "P_Nom",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_BOISTYPES"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_BOISTYPES"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_BOISTYPES"
                }
              }
            }
          }
        }
      }
    },
    "/api/BoisType/AddBoisType": {
      "post": {
        "tags": [
          "BoisType"
        ],
        "operationId": "AddBoisType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_BOISTYPES"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_BOISTYPES"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_BOISTYPES"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/BoisType/UpdateBoisType": {
      "put": {
        "tags": [
          "BoisType"
        ],
        "operationId": "UpdateBoisType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_BOISTYPES"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_BOISTYPES"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_BOISTYPES"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BoisType/DeleteBoisType": {
      "delete": {
        "tags": [
          "BoisType"
        ],
        "operationId": "DeleteBoisType",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "C_AUTHENTIFICATION": {
        "type": "object",
        "properties": {
          "idAuth": {
            "type": "integer",
            "format": "int32"
          },
          "adresseMailAuth": {
            "type": "string",
            "nullable": true
          },
          "passwordAuth": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "C_BOISTYPES": {
        "type": "object",
        "properties": {
          "idBois": {
            "type": "integer",
            "format": "int32"
          },
          "nomBois": {
            "type": "string",
            "nullable": true
          },
          "descriptionBois": {
            "type": "string",
            "nullable": true
          },
          "dispoBois": {
            "type": "boolean"
          },
          "prix": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "C_HASHMDP": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "hachage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}